Add exports to barrel file, clean up file structure
JIRA: RAIL-2847
