+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/KpiAlerts/KpiAlertDialog/KpiAlertDialogBrokenFilters/utils/filterUtils.ts
@@ -1,0 +1,30 @@
+// (C) 2007-2021 GoodData Corporation
+import { IBrokenAlertFilter, isBrokenAlertAttributeFilter } from "../../../types";
+
+interface ILabelFilterData {
+    title: string;
+    selection: string;
+    isDate: boolean;
+    isAllSelected: boolean;
+    selectionSize: number;
+}
+
+export function getFilterLabelFilter(item: IBrokenAlertFilter): ILabelFilterData {
+    if (isBrokenAlertAttributeFilter(item)) {
+        return {
+            isAllSelected: item.isAllSelected,
+            isDate: false,
+            selection: item.selection,
+            selectionSize: item.selectionSize,
+            title: item.title,
+        };
+    } else {
+        return {
+            isAllSelected: false,
+            isDate: true,
+            selection: item.dateFilterTitle,
+            selectionSize: null,
+            title: item.title,
+        };
+    }
+}
