+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/KpiAlerts/KpiAlertDialog/utils/filterUtils.ts
@@ -24,36 +24,6 @@ import isEqual from "lodash/isEqual";
 import last from "lodash/last";
 import sortBy from "lodash/sortBy";
 
-import { IBrokenAlertFilter, isBrokenAlertAttributeFilter } from "../types";
-
-interface ILabelFilterData {
-    title: string;
-    selection: string;
-    isDate: boolean;
-    isAllSelected: boolean;
-    selectionSize: number;
-}
-
-export function getFilterLabelFilter(item: IBrokenAlertFilter): ILabelFilterData {
-    if (isBrokenAlertAttributeFilter(item)) {
-        return {
-            isAllSelected: item.isAllSelected,
-            isDate: false,
-            selection: item.selection,
-            selectionSize: item.selectionSize,
-            title: item.title,
-        };
-    } else {
-        return {
-            isAllSelected: false,
-            isDate: true,
-            selection: item.dateFilterTitle,
-            selectionSize: null,
-            title: item.title,
-        };
-    }
-}
-
 export function isKpiAlertDateFilterSameAsDashboard(
     alert: IWidgetAlertDefinition,
     appliedFilters: IFilter[],
