+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/KpiAlerts/index.ts
@@ -1,5 +1,19 @@
-// (C) 2007-2020 GoodData Corporation
+// (C) 2007-2021 GoodData Corporation
 
 import { DashboardItemWithKpiAlert } from "./DashboardItemWithKpiAlert";
+import KpiAlertDialog from "./KpiAlertDialog/KpiAlertDialog";
+import {
+    IAttributeFilterMetaCollection,
+    IBrokenAlertFilterBasicInfo,
+    enrichBrokenAlertsInfo,
+    getBrokenAlertFiltersBasicInfo,
+} from "./utils/brokenFilterUtils";
 
-export { DashboardItemWithKpiAlert };
+export {
+    DashboardItemWithKpiAlert,
+    KpiAlertDialog,
+    IAttributeFilterMetaCollection,
+    IBrokenAlertFilterBasicInfo,
+    enrichBrokenAlertsInfo,
+    getBrokenAlertFiltersBasicInfo,
+};
