+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/KpiAlerts/utils/translationUtils.ts
